[Cogs/Respects] Add support for replies
`Respects.payRespects` now queries the last respect (if any, else `None`)  to obtain its reference (i.e., the message to which the respect was paid), and attaches it (the reference) to the new respect message.

`Respects.checkLastRespect` now queries the last respect (if any, else `None`) and checks if the last respect and the current respect (i.e., at the time of invocation of the function) reply to two different messages. If yes, the current respect chain's data (including the user list) gets reset so as to prepare for a new respect chain.