[Tags] Save data in the config directory
Make `Tags` write to the directory specified in `Config` class (which is
specified by the cog to be the cog data path set up as a part of bot
instance setup) which is writable.

Currently the cog writes to a file whose path is specified as a purely
relative path, which, at absolute path resolution time, will become a
path under wherever the bot is running (most likely not the writable cog
data path) and the cog will fail when it does not have write permission
to the said resolved path.

This commit specifies a path relative to the cog data path which is
writable by cogs, which avoids the aforementioned issue.