[CI] Use upstream test pipeline for cog unit test
It appears that the newest version of `pytest-asyncio` does not provide
the `event_loop` Pytest fixture anymore.

A simple way to let cog unit tests still run with that fixture is to
gate the version of the said dependency.

Because upstream's requirement list already does that, this commit
removes the GitHub workflow file `cogUnitTests.yml` and simply extends
the Pytest test paths (in `pyproject.toml`) to achieve the same effect
while using upstream's CI configuration.